[breaking changes] Make functions in onlinejudge_template/generator/__init__.py private
They were public and written our documents unintentionally.
